pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "Sloths_Pain"
version = "0.2.2"
description = "Python project using PyTorch and PyTorch Lightning."
authors = [
    { name = "Odei and Lucas" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"

dependencies = [
    "torch==2.8.0",
    "torchvision>=0.20.0",
    "numpy==2.3.3",
    "pandas==2.3.3",
    "pillow>=10.4.0",
    "matplotlib==3.10.6",
    "seaborn==0.13.2",
    "pytorch-lightning==2.5.5",
    "scikit-learn>=1.5.0",
    "pytest==8.4.2",
    "gradio==5.49.1"
]

[project.optional-dependencies]
dev = ["pytest>=8.3.3", "black>=24.10.0", "flake8>=7.1.1"]

[tool.setuptools.packages.find]
where = ["src"]
include = ["sloths*"]
namespaces = false

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.pytest.ini_options]
testpaths = ["tests"]

